move launchers to subpackage
